+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [1.2.0](https://github.com/cedx/html.net/compare/v1.1.0...v1.2.0)
-- Added the `Out-String` cmdlet.
+- Added the `Out-File` and `Out-String` cmdlets.
 
 ## Version [1.1.0](https://github.com/cedx/html.net/compare/v1.0.1...v1.1.0)
 - Added support for `[switch]` parameters in attribute values.